Première version fonctionnant avec le tutorial de correspondance grouping

+++ stalker/include/stalker/Shape3D.hpp
@@ -36,7 +36,7 @@ class Shape{
 	
 	Shape(const std::string& name, double sampling_size) : _id(name), _shape(new pcl::PointCloud<T>()),_shape_keypoints(new pcl::PointCloud<T>()),_shape_normals(new pcl::PointCloud<NormalType>()), _desc(new pcl::PointCloud<DescriptorType>()), _descrRad(0.02), _shape_ss (sampling_size){};
 	
-	Shape(const std::string& name, double sampling_size, int descriptor_radius) : _id(name), _shape(new pcl::PointCloud<T>()),_shape_keypoints(new pcl::PointCloud<T>()),_shape_normals(new pcl::PointCloud<NormalType>()), _desc(new pcl::PointCloud<DescriptorType>()), _descrRad(descriptor_radius), _shape_ss (sampling_size){};
+	Shape(const std::string& name, double sampling_size, double descriptor_radius) : _id(name), _shape(new pcl::PointCloud<T>()),_shape_keypoints(new pcl::PointCloud<T>()),_shape_normals(new pcl::PointCloud<NormalType>()), _desc(new pcl::PointCloud<DescriptorType>()), _descrRad(descriptor_radius), _shape_ss (sampling_size){};
 
 	virtual ~Shape(){
 		std::cout<<"delete shape "<<_id <<std::endl;
